+++ libs/forms/src/lib/declarative/form-fields/specialized/index.ts
@@ -15,6 +15,7 @@ export {
 export { FieldAddress } from './field-address'
 export { FieldCity } from './field-city'
 export { type ColorPickerFieldProps, FieldColorPicker } from './field-color-picker'
+export { FieldEditIntent } from './field-edit-intent'
 export { FieldFileUpload, type FileUploadFieldProps } from './field-file-upload'
 export { FieldOTPInput } from './field-otp-input'
 export { FieldPhone } from './field-phone'